Starting work on a no-diode mono-PCB layout using Graph Theory
See issue #1 and https://astrobeano.blogspot.com/2025/05/ergo-mech-keyboard-wiring-using-tutte-coxeter-graph.html

The plan is to design an alternative larger single PCB compatible with the original Gamma-Omega case.
This will use the same class of 26 GPIO microcontroller and a 26 of 30 vertex subset of the bipartite Tutte-Coxeter graph to control 36 keys (graph edges) within a sparse 13x13 virtual scanning matrix.