Switch from half header to full Raspberry Pi Pico footprint.
Using a footprint submitted to erogegen as pull request 144.

Note: Will need to move the cutting edge up a couple of mm as
I want to use GP zero (unused in split PCB design). There looks
to be enough clearance with the original case design.

This is setup with the hotswaps on the back (default) and
controller also on the back (although it will actually be upside
down on the front with header pins for spacing). i.e. once
assembled the rp2040 will be face down, and the PCB front
silkscreen will be visible along with the keycaps. The hotswap
sockets and silkscreen art on the PCB back will face the table.