Bundle third-party Raspberry Pi Pico footprint
This is from Ergogen pull request 144 by Vladimir Andral in Dec 2024
(GitHub user VladAndral), based on a first draft by Thomas Andersen
in Feb 2022 (GitHub user TTA777) here:

https://github.com/TTA777/ergogen/commit/817e013d4880876b9fd51707de45c7b36a1b5077

This is therefore (like Ergogen) under the MIT License. If this
or something like it is added to Ergogen, we can drop this.