Redo the top edge to make a USB port cutout.
The central bottom edge rectangle is now taller and gives the bottom
edge of the USB cut out. This remains anchored to the thumbs.

The top edge is now two mirrored rectangles, and provides the sides
of the USB cutout. This is now anchored top lift to the top pinky.