Traces and ground plane in KiCad v9.0.2
This was the result of a lot of iterative work, which does not
work well in version control (too much floating point noise).

This passes the Design Rules Checker except for a warning from
each ceoloide footprint not being in the library, a warning for
the USB cut-out clipping the controller footprint (deliberate),
and 2 errors about not enough spokes for ground plane zones
(cofigurable threshold).