Gamma Omega 'Hesse' is a better name than placeholder 'GT36K'.
The underlying graph of the affine plane of order 3 is also known as the
Hesse Configuration <https://en.wikipedia.org/wiki/Hesse_configuration>.
So what we are using could reasonably be called the 'Hesse Configuration
Incidence Graph'. A visually pleasing drawing of this is tricky, but the
Hesse Configuration Graph would work as a silk screen illustration.