Add battery cut-out, eg 3.7v 110mAh 301230
That is a commonly used size of 3mm thickness, and 12mm by 30mm footprint.
This placement aims to leave a keepout at the bottom of the controller for
the bluetooth signal, but above the case taper and structural cross beam.
Will still need to adust the top and or bottom case for this...

This location should leave enough space either side of the battery space
for traces similar to the layout used on the TC36K.

I am currently thinking a battery with the JST socket, and wiring a matching
JST connector with wires to the battery terminals at the top of the controller
(by the USB port), possibly via an on/off switch. Running thick traces to this
battery location may be too tricky with the switch wiring in the way?