Bottom case for 'Hesse' Bluetooth variant (joint effort)
Based on the original covered bottom case with the following changes:

* Simlified the hotswap cutouts no covered (lighter this way).
* Added 1.4mm battery resess cutout
* Added cutout for the JST PH 2.0 battery connector legs
* Added cutout for the reset button legs
* Added cutout for the Nice!Nano v2 or clone middle pins

This retains the full length header cutouts for the Raspberry Pi Pico,
so should still work with the original or TC36K PCB too.

Done in a mixture of tools, initially editing the original case in
FreeCAD (peterjc) which made the file much larger. Imported back
into AF, glitches cleaned up, and re-exported as a smaller STEP file
(unspecworks).